Use shell command to pipe ACCEPT to EULA
In newer releases of Vertica, when admintools is run the first time, it expects the user to manually accept the EULA. It is not possible to do this with the  Ansible "command". Switching to using the Ansible "shell" allows ACCEPT to be piped to admintools when start-up the first time.